Abstract atv document from its model
This commit adds an abstraction layer between the atv document that is manipulated by the user and its actual representation in a file.
It does _not_ contain any revisions or model changes, yet.

Background:
Simple values can be encoded in just one line or as a more complex structure containing additional metadata (access, uuid, ...).
Both representations in an ATV file are valid and can express the same value, but with a different encoding
=> an additional layer is needed that takes care of unifying the access to the atv document model.